requirements.txt
pygame
pillow
tk